Correct test for rvalue bind(unique_ptr, f).
The rvalue overload of monad<std::unique_ptr<T>>::bind(ptr,f) requires that T
has the same type as f(*ptr). In this case, f returns a float, but got supplied
an int.

+++ tests/memory_tests.cpp
@@ -248,7 +248,7 @@ test_set memory_tests{
 
 				return *(p >>= f) == .5f
 					// Test the rvalue overload.
-					&& *(std::unique_ptr<int>(new int(1)) >>= f) == .5f
+					&& *(std::unique_ptr<float>(new float(1)) >>= f) == .5f
 					&& (null >>= f) == nullptr;
 			})
 		),
